backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
PyJWT==2.8.0
requests==2.31.0
google-cloud-speech==2.20.0        # optional, only if you want real transcription
llama-index==0.9.38               # adjust to your installed version (adapter code written generically)
